Added tests and example
